# specs/openapi/api.v1.yaml
openapi: "3.1.0"
info:
  title: "LongoX Platform API"
  version: "1.0.0"
  description: "LongoX API — low-code automation and internal tools platform"
  contact:
    name: "LongoX Support"
    url: "https://longox.io/support"
  license:
    name: "Proprietary"
servers:
  - url: "https://api.longox.io"
    description: "Production"
  - url: "https://staging-api.longox.io"
    description: "Staging"
  - url: "http://localhost:3000"
    description: "Development"
paths:
...
  /api/v1/connectors/marketplace:
    get:
      operationId: listConnectorMarketplace
      tags: ["Connectors"]
      summary: "List available connectors in marketplace"
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PaginationPage"
        - $ref: "#/components/parameters/PaginationLimit"
        - $ref: "#/components/parameters/ConnectorCategoryFilter"
        - $ref: "#/components/parameters/ConnectorSearch"
      responses:
        "200":
          description: "Connector marketplace listing"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Connector"
                  total:
                    type: integer
...
  /api/v1/connectors/{id}/actions:
    get:
      operationId: listConnectorActions
      tags: ["Connectors"]
      summary: "List connector actions"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Connector actions"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConnectorAction"
  /api/v1/connectors/{id}/triggers:
    get:
      operationId: listConnectorTriggers
      tags: ["Connectors"]
      summary: "List connector triggers"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Connector triggers"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConnectorTrigger"
  /api/v1/connectors/{id}/versions:
    get:
      operationId: listConnectorVersions
      tags: ["Connectors"]
      summary: "List connector versions"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Connector versions"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConnectorVersion"
  /api/v1/connectors/{id}/install:
    post:
      operationId: installConnector
      tags: ["Connectors"]
      summary: "Install a connector"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Installed connector"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectorInstallation"
  /api/v1/connectors/installations:
    get:
      operationId: listInstallations
      tags: ["Connectors"]
      summary: "List tenant connector installations"
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      responses:
        "200":
          description: "Installations"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConnectorInstallation"
  /api/v1/connectors/installations/{id}:
    delete:
      operationId: uninstallConnector
      tags: ["Connectors"]
      summary: "Uninstall a connector"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: "Uninstalled"
  /api/v1/connectors/installations/{id}/configure:
    post:
      operationId: configureConnector
      tags: ["Connectors"]
      summary: "Configure a connector installation"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                config:
                  type: object
      responses:
        "200":
          description: "Configured installation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectorInstallation"
  /api/v1/connectors/installations/{id}/execute:
    post:
      operationId: executeConnectorAction
      tags: ["Connectors"]
      summary: "Execute a connector action"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteConnectorInput"
      responses:
        "200":
          description: "Action result"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectorExecuteResult"
  /api/v1/connectors/installations/{id}/test:
    post:
      operationId: testConnector
      tags: ["Connectors"]
      summary: "Test connector connection"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Test result"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectorTestResult"
  /api/v1/connectors/installations/{id}/upgrade:
    post:
      operationId: upgradeConnector
      tags: ["Connectors"]
      summary: "Upgrade connector version"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                versionId:
                  type: string
      responses:
        "200":
          description: "Upgraded installation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectorInstallation"
  /api/v1/connectors/installations/{id}/rollback:
    post:
      operationId: rollbackConnector
      tags: ["Connectors"]
      summary: "Rollback connector to previous version"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Rollback result"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectorRollbackResult"
  /api/v1/connectors/installations/{id}/webhooks:
    post:
      operationId: registerConnectorWebhook
      tags: ["Connectors"]
      summary: "Register webhook for connector"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookRegistrationInput"
      responses:
        "200":
          description: "Webhook registered"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookRegistration"
  /api/v1/connectors/installations/{id}/poll:
    post:
      operationId: pollConnector
      tags: ["Connectors"]
      summary: "Poll connector for new events"
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                triggerId:
                  type: string
      responses:
        "200":
          description: "Poll result"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PollResult"

  # ─── AI ─────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Bearer token obtained from /auth/login or /auth/register"
  parameters:
    TenantHeader:
      name: x-tenant-id
      in: header
      schema:
        type: string
      description: "Tenant context for multi-tenant requests"
    PaginationPage:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PaginationLimit:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    ConnectorCategoryFilter:
      name: category
      in: query
      schema:
        type: string
    ConnectorSearch:
      name: search
      in: query
      schema:
        type: string
...
  schemas:
...
    Connector:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        displayName:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        category:
          type: string
        trustLevel:
          type: string
        scope:
          type: string
          nullable: true
        authType:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ConnectorAction:
      type: object
      properties:
        id:
          type: string
        connectorId:
          type: string
        actionId:
          type: string
        name:
          type: string
        description:
          type: string
        inputSchema:
          type: object
        outputSchema:
          type: object
    ConnectorTrigger:
      type: object
      properties:
        id:
          type: string
        connectorId:
          type: string
        triggerId:
          type: string
        triggerType:
          type: string
        name:
          type: string
        description:
          type: string
        config:
          type: object
        pollingInterval:
          type: integer
          nullable: true
    ConnectorVersion:
      type: object
      properties:
        id:
          type: string
        connectorId:
          type: string
        semver:
          type: string
        manifestJson:
          type: object
        artifactRef:
          type: string
          nullable: true
        signatureRef:
          type: string
          nullable: true
        isSupported:
          type: boolean
        publishedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    ConnectorInstallation:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        connectorVersionId:
          type: string
        status:
          type: string
        installedAt:
          type: string
          format: date-time
        installedBy:
          type: string
          nullable: true
        config:
          type: object
        createdAt:
          type: string
          format: date-time
    ExecuteConnectorInput:
      type: object
      required: [actionId]
      properties:
        actionId:
          type: string
        input:
          type: object
    ConnectorExecuteResult:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          type: string
          nullable: true
    ConnectorTestResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        latencyMs:
          type: integer
    ConnectorRollbackResult:
      type: object
      properties:
        success:
          type: boolean
        previousVersion:
          type: string
        currentVersion:
          type: string
    WebhookRegistrationInput:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        secret:
          type: string
    WebhookRegistration:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
        events:
          type: array
          items:
            type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    PollResult:
      type: object
      properties:
        hasNewEvents:
          type: boolean
        eventCount:
          type: integer
        lastCursor:
          type: string
...